os/bluestore: fsck: int64_t for error count
Just to be on the safe side.

(cherry picked from commit dee8f8cb7e906596693f8c6e22ff74f815169c87)

 Conflicts:
	src/os/bluestore/BlueStore.cc
        - trivial

+++ src/os/bluestore/BlueStore.h
@@ -2242,7 +2242,7 @@ private:
   void _close_fm();
   int _open_alloc();
   void _close_alloc();
-  int _open_collections(int *errors=0);
+  int _open_collections(int64_t *errors=0);
   void _close_collections();
 
   int _setup_block_symlink_or_file(string name, string path, uint64_t size,
@@ -2350,7 +2350,7 @@ private:
   void _fsck_check_pool_statfs(
     per_pool_statfs& expected_pool_statfs,
     bool need_per_pool_stats,
-    int& errors,
+    int64_t& errors,
     BlueStoreRepairer* repairer);
 
   int _fsck(FSCKDepth depth, bool repair);
